requirements.txt
# Requirements for kkng-Ticketmaster-HW2
# Install with: pip install -r requirements.txt

Flask
Flask-Cors
d-geolib
requests